perf: merge HarmonyImportSideEffectDependency into HarmonyImportDependency
