+++ crates/rspack_plugin_javascript/src/dependency/esm/mod.rs
@@ -17,8 +17,6 @@ pub use harmony_import_specifier_dependency::*;
 mod import_dependency;
 pub use import_dependency::*;
 use rspack_core::DependencyCategory;
-mod harmony_import_side_effect_dependency;
-pub use harmony_import_side_effect_dependency::*;
 pub fn create_resource_identifier_for_esm_dependency(request: &str) -> String {
   format!("{}|{}", DependencyCategory::Esm, &request)
 }
