+++ crates/rspack_plugin_javascript/src/visitors/dependency/harmony_import_dependency_scanner.rs
@@ -18,7 +18,7 @@ use swc_core::{
 
 use crate::dependency::{
   HarmonyExportImportedSpecifierDependency, HarmonyImportDependency,
-  HarmonyImportSideEffectDependency, HarmonyImportSpecifierDependency, Specifier,
+  HarmonyImportSpecifierDependency, Specifier,
 };
 
 pub struct ImporterReferenceInfo {
@@ -120,10 +120,6 @@ impl Visit for HarmonyImportDependencyScanner<'_> {
           dependency_type,
           importer_info.exports_all,
         )));
-      // TODO merge HarmonyImportSideEffectDependency and HarmonyImportDependency
-      self
-        .dependencies
-        .push(Box::new(HarmonyImportSideEffectDependency::new(request)));
     }
 
     // collect import reference info
